feat: move activity interfaces to separate file
